BeansBot Update V13
I'm rewriting the beansbot for discord v13. The code for the previous version is in ./deprecated.

I want to challenge myself to use app commands this time and write things more clearly. I also deleted the previous repository for this project because it wasn't done correctly and I'm making this a portfolio piece.

This first commit has a simple setup with only #confessions working. I will add more functionality soon.